Add Jan's isolation ring sums to value map
Also remove firstLayerEnergy per Florian's comment
Also kill all evil carriage returns in Jan's files
